+++ src/include/sof/panic.h
@@ -31,71 +31,16 @@
 #ifndef __INCLUDE_SOF_IPC_PANIC__
 #define __INCLUDE_SOF_IPC_PANIC__
 
-#include <sof/sof.h>
-#include <sof/alloc.h>
-#include <sof/mailbox.h>
-#include <sof/interrupt.h>
-#include <sof/trace.h>
-#include <platform/platform.h>
-#include <uapi/ipc/trace.h>
 #include <stdint.h>
-#include <stdlib.h>
 
-static inline void dump_panicinfo(void *addr,
-				  struct sof_ipc_panic_info *panic_info)
-{
-	if (!panic_info)
-		return;
-	rmemcpy(addr, panic_info, sizeof(struct sof_ipc_panic_info));
-	dcache_writeback_region(addr, sizeof(struct sof_ipc_panic_info));
-}
+struct sof_ipc_panic_info;
 
-/* panic and rewind stack */
-static inline void panic_rewind(uint32_t p, uint32_t stack_rewind_frames,
-				struct sof_ipc_panic_info *panic_info)
-{
-	void *ext_offset;
-	size_t count;
-	uint32_t oldps;
-
-	/* disable all IRQs */
-	oldps = interrupt_global_disable();
-
-	/* dump DSP core registers */
-	ext_offset = arch_dump_regs(oldps);
-
-	/* dump panic info, filename ane linenum */
-	dump_panicinfo(ext_offset, panic_info);
-	ext_offset += sizeof(struct sof_ipc_panic_info);
-
-	count = MAILBOX_EXCEPTION_SIZE -
-		(size_t)(ext_offset - mailbox_get_exception_base());
-	/* dump stack frames */
-	p = dump_stack(p, ext_offset, stack_rewind_frames, count);
-
-	/* panic - send IPC oops message to host */
-	platform_panic(p);
-
-	/* flush last trace messages */
-	trace_flush();
-
-	/* and loop forever */
-	while (1) {};
-}
+void dump_panicinfo(void *addr, struct sof_ipc_panic_info *panic_info);
+void panic_rewind(uint32_t p, uint32_t stack_rewind_frames,
+		  struct sof_ipc_panic_info *panic_info);
+void __panic(uint32_t p, char *filename, uint32_t linenum);
 
 /* panic dump filename and linenumber of the call */
 #define panic(x) __panic((x), (__FILE__), (__LINE__))
-
-static void __panic(uint32_t p, char *filename, uint32_t linenum)
-{
-	struct sof_ipc_panic_info panicinfo;
-	int strlen;
-
-	strlen = rstrlen(filename);
-	panicinfo.linenum = linenum;
-	rmemcpy(panicinfo.filename, filename, strlen + 1);
-
-	panic_rewind(p, 0, &panicinfo);
-}
 
 #endif
